Move make_unclosed_delims_error to lexer/diagonostics.rs
